+++ src/main/java/com/mapshot/api/domain/notice/NoticeListResponse.java
@@ -1,6 +1,5 @@
-package com.mapshot.api.presentation.notice.model;
+package com.mapshot.api.domain.notice;
 
-import com.mapshot.api.domain.notice.NoticeEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,21 +11,18 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class NoticeDetailResponse {
+public class NoticeListResponse {
 
     private long id;
     private String noticeType;
     private String title;
-    private String content;
     private LocalDateTime createdDate;
 
-
-    public static NoticeDetailResponse fromEntity(NoticeEntity noticeEntity) {
-        return NoticeDetailResponse.builder()
+    public static NoticeListResponse fromEntity(NoticeEntity noticeEntity) {
+        return NoticeListResponse.builder()
                 .id(noticeEntity.getId())
                 .noticeType(noticeEntity.getNoticeType().getKorean())
                 .title(noticeEntity.getTitle())
-                .content(noticeEntity.getContent())
                 .createdDate(noticeEntity.getCreatedDate())
                 .build();
     }
