+++ src/main/java/com/mapshot/api/presentation/notice/NoticeController.java
@@ -1,11 +1,11 @@
 package com.mapshot.api.presentation.notice;
 
 
+import com.mapshot.api.domain.notice.NoticeDetailResponse;
+import com.mapshot.api.domain.notice.NoticeListResponse;
 import com.mapshot.api.domain.notice.NoticeService;
 import com.mapshot.api.infra.auth.annotation.PreAuth;
 import com.mapshot.api.infra.auth.enums.Accessible;
-import com.mapshot.api.presentation.notice.model.NoticeDetailResponse;
-import com.mapshot.api.presentation.notice.model.NoticeListResponse;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.RequiredArgsConstructor;
