+++ src/test/java/com/mapshot/api/domain/notice/NoticeServiceTest.java
@@ -2,8 +2,6 @@ package com.mapshot.api.domain.notice;
 
 import com.mapshot.api.infra.exception.ApiException;
 import com.mapshot.api.infra.exception.status.ErrorCode;
-import com.mapshot.api.presentation.notice.model.NoticeDetailResponse;
-import com.mapshot.api.presentation.notice.model.NoticeListResponse;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -43,7 +41,7 @@ class NoticeServiceTest {
     void release() {
         noticeRepository.deleteAll();
     }
-    
+
 
     @Test
     void 시작_id를_0으로_전송_시_가장_최근_게시글부터_10개_반환() {
