+++ src/test/java/com/mapshot/api/presentation/notice/NoticeControllerTest.java
@@ -3,13 +3,8 @@ package com.mapshot.api.presentation.notice;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mapshot.api.SlackMockExtension;
-import com.mapshot.api.domain.notice.NoticeEntity;
-import com.mapshot.api.domain.notice.NoticeRepository;
-import com.mapshot.api.domain.notice.NoticeService;
-import com.mapshot.api.domain.notice.NoticeType;
+import com.mapshot.api.domain.notice.*;
 import com.mapshot.api.infra.auth.Validation;
-import com.mapshot.api.presentation.notice.model.NoticeDetailResponse;
-import com.mapshot.api.presentation.notice.model.NoticeListResponse;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
